requirements.txt
# Django y REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3

# Imágenes
Pillow==10.1.0

# Base de datos
mysqlclient==2.2.0

# Autenticación JWT
djangorestframework-simplejwt==5.3.0

# Celery y Message Broker
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Django Channels (WebSockets para señalización digital)
channels==4.0.0
channels-redis==4.1.0

# Variables de entorno
python-dotenv==1.0.0

# Servidores de Producción
gunicorn==21.2.0
daphne==4.0.0

# Utilidades
python-dateutil==2.8.2
pytz==2023.3

# Validación y serialización
pydantic==2.5.0
